+++ CHANGELOG.md
@@ -12,6 +12,8 @@ interface to `module:oauth2orize.RequestProcessor`.
 interface to `module:oauth2orize.RequestParametersProcessor`.
 - Renamed `http://i.authnomicon.org/oauth2/authorization/http/ResponseParameters`
 to `module:oauth2orize.responseParametersFn`.
+- Renamed `http://i.authnomicon.org/oauth2/token/http/ResponseParameters`
+to `module:@authnomicon/oauth2.tokenResponseParametersFn`.
 
 ## [0.0.4] - 2021-12-15
 ### Added